fix(windows): restore wifi logic
resolves #3049
